Python/requirements.txt
flask
flask_csv
mysql-connector-python
